requirements.txt
# Core
streamlit>=1.35.0
pandas>=2.2.0
numpy>=1.26.0

# PDF processing
pdfplumber>=0.11.0
pypdf>=4.2.0

# LLM (Phase 1: Claude API for unstructured analysis)
anthropic>=0.28.0

# Data quality utilities
great-expectations>=0.18.0   # optional — used for advanced rule sets
phonenumbers>=8.13.0         # PII: phone number detection
presidio-analyzer>=2.2.354   # PII detection (names, SSN, emails, etc.)
presidio-anonymizer>=2.2.354

# Reporting
jinja2>=3.1.4
openpyxl>=3.1.2
python-docx>=1.1.2

# Utilities
pyyaml>=6.0.1
python-dotenv>=1.0.1
loguru>=0.7.2
rich>=13.7.1

# Viz
plotly>=5.22.0
altair>=5.3.0
